1、修复使用ESP32 release1.0.0后，产生的warnings 2、将蓝牙/串口接受发送缓存加大，以防止出现越界读写，改为Rx_Data[16];Tx_Data[16]; 3、修复静态显示彩灯更新异常 4、修复IIC读的过程，判断Wire.available()前 先执行判断 Wire.requestFrom的返回是否正确

+++ XT1511_I2C/XT1511_I2C.h
@@ -67,7 +67,7 @@ class XT1511_I2C
     void LED_Flush(void);
 
   private:
-      int _device_address = 0x11;
+      byte _device_address = 0x11;
       byte LEDs_Data[36];
       byte LEDs_DataResult[36];
